refactor: organize code into token function
