+++ src/token.ts
@@ -1,7 +1,8 @@
 import { Address } from '@graphprotocol/graph-ts'
 import { Contract } from './types/Contract/Contract'
 import { Token } from './types/schema'
-import { CONTRACT_ADDRESS, DECIMAL_ZERO, INT_ZERO } from "./constants"
+import { convertTokenToDecimal } from "./helpers"
+import { BURN_ADDRESS, CONTRACT_ADDRESS, DECIMAL_ZERO, INT_ZERO } from "./constants"
 
 // Gets the singleton instance of the token entity
 export function getTokenInstance() : Token {
@@ -26,4 +27,21 @@ export function getTokenInstance() : Token {
   }
 
   return token as Token
+}
+
+export function updateTokenSupply() : void {
+  let contract = Contract.bind(Address.fromString(CONTRACT_ADDRESS))
+  let token = getTokenInstance()
+
+  // Get the amount burned
+  let burnAddress = Address.fromString(BURN_ADDRESS)
+  let totalBurned = convertTokenToDecimal(contract.balanceOf(burnAddress))
+
+  // Determine remaining supply from burned
+  token.totalBurned = totalBurned
+  token.remainingSupply = token.totalSupply.minus(totalBurned)
+
+  // Update total fees
+  token.totalFees = convertTokenToDecimal(contract.totalFees())
+  token.save()
 }
